examples/requirements.txt
tiny-proxy>=0.2.0
PyYAML>=3.12